fix(runtime): split plan-to-git hook templates
